Tweaks.
Reduced fire sound distance config value.
Reduced minimal grenade speed to produce bounce sounds.
Touched up the modular stock textures.
Overhauled the slide on automatic pistol.
Reloading now stops and restricts sprinting.
The gun is no longer raised when trying to shoot with a broken and loaded gun.
Recoil adjustments.
Removed several Util classes.

+++ src/main/java/com/mrcrayfish/guns/client/render/crosshair/DynamicCrosshair.java
@@ -76,7 +76,7 @@ public class DynamicCrosshair extends Crosshair
             	GunItem gun = (GunItem) heldItem.getItem();
             	Gun modifiedGun = gun.getModifiedGun(heldItem);
             	float aiming = (float) AimingHandler.get().getNormalisedAdsProgress();
-                float sprintTransition = (float) GunRenderingHandler.get().getSprintTransition(Minecraft.getInstance().getFrameTime());
+                float sprintTransition = GunRenderingHandler.get().getSprintTransition(Minecraft.getInstance().getFrameTime());
                 float spreadCount = (SpreadTracker.get(mc.player).getNextSpread(gun,aiming));
             	float spreadModifier = ((spreadCount+(1F/Math.max(Config.COMMON.maxCount.get(),1F)))*Math.min(Mth.lerp(partialTicks, this.prevFireBloom, this.fireBloom),1F));
             	spreadModifier = (float) (Mth.lerp(sprintTransition*0.5, spreadModifier, 1.0F));
